fix: created advance modal and find friend modal
