+++ src/pages/Friend/MyFriend.tsx
@@ -2,6 +2,7 @@ import WidthSizeDetection from "../../assets/config/WidthSizeDetection";
 import Nav from "../../components/Nav/Nav";
 import { RoundedButton } from "../../components/button/Button";
 import Link from "../../components/link/Link";
+import AdvanceModalWithBtn from "../../components/modal/AdvanceModalWithButton";
 import RowStandardModal from "../../components/modal/RowStandardModal";
 
 const MyFriend = () => {
@@ -19,9 +20,11 @@ const MyFriend = () => {
         <div
           className={`${widthSize.mediumDevice ? "basis-9/12" : "basis-7/12"}`}
         >
-          <RoundedButton
-            text="Find friends"
+          <AdvanceModalWithBtn
+            modalOpenBtnName="Find friends"
+            modalTitle="Find friends"
             className="bg-white h-10 ml-4 mb-5"
+            isShowFindFriendModal={true}
           />
           <p className="ml-5 mb-5 text-2xl font-bold">
             You haven't follow any friend yet.
